add register script

+++ clean.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
 
-rm -rf cas-server
-rm -rf strapi
+# clean strapi DB
+rm -rf strapi/examples/getstarted/.tmp
+echo "Strapi DB deleted"
+if [[ "$1" == "all" ]] ; then
+  rm -rf cas-server
+  rm -rf strapi
+else
+  echo "Run \"$0 all\" to delete CAS overlay and Strapi cloned repo"
+fi
